market-maker/requirements.txt
websockets
httpx
cryptography
boto3
pyarrow
polars
python-dotenv
